Image Classification/CGIAR Wheat Growth Stage Challenge/Nuno/competition_CGIAR_user_ngcferreira_3rd_place/requirements.txt
absl-py==0.9.0
albumentations==0.4.6
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.3
idna==2.8
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.3
joblib==0.17.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jupyterlab-pygments==0.1.2
kaggle==1.5.6
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
mistune==0.8.4
more-itertools==7.2.0
nbclient==0.5.0
nbconvert==6.0.7
nbformat==4.4.0
nest-asyncio==1.4.1
networkx==2.4
notebook==6.0.2
numpy==1.19.2
opencv-python==4.4.0.44
pandas==1.1.3
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.2.0
pipenv==2018.11.26
prometheus-client==0.7.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
pyparsing==2.4.7
pyrsistent==0.15.5
python-dateutil==2.8.1
python-slugify==4.0.0
pytorchcv==0.0.58
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
scikit-image==0.17.2
scikit-learn==0.23.2
scikit-multilearn==0.2.0
scipy==1.5.2
Send2Trash==1.5.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
smmap==3.0.4
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tf-slim==1.2.0
threadpoolctl==2.1.0
tifffile==2020.7.4
torch==1.6.0
tornado==6.0.3
tqdm==4.39.0
traitlets==4.3.3
urllib3==1.24.3
virtualenv==16.7.7
virtualenv-clone==0.5.3
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==0.6.0
